pyproject.toml
[project]
name = "webwallpaper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pywebview>=6.0",
    "pywin32>=311",
    "pystray>=0.19",
    "pillow>=10.0",
    "pyinstaller>=6.0",
    "psutil>=5.9",
]
